+++ src/main/java/microservice/bien/serviceImpl/Gestion_Users/Agent_ImmobilierServiceImpl.java
@@ -21,7 +21,7 @@ public class Agent_ImmobilierServiceImpl implements Agent_ImmobilierService {
     }
 
     @Override
-    public Agent_Immobilier getById(Integer id) {
+    public Agent_Immobilier getById(Long id) {
         return this.agent_immobilierDao.getOne(id);
     }
